fix(docs): move DBModule to separate file for Compodoc compatibility
Compodoc wasn't recognizing the @Module decorator when declared inline in index.ts.
I extracted DBModule into db.module.ts and re-exported it to preserve existing imports.
