+++ src/shared/schema/index.ts
@@ -1,15 +1,13 @@
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { DataLog, DataLogSchema } from './data.log.schema';
-import { configs } from '../configs';
+import { ConfigService } from '@nestjs/config';
 import { Connection, ConnectOptions, createConnection } from 'mongoose';
-import { Module } from '@nestjs/common';
-import { User, UserSchema } from './user.schema';
+import { DataLog, DataLogSchema } from './data.log.schema';
 import { EventSchema } from './events.schema';
+import { User, UserSchema } from './user.schema';
 
 // All Schema Models
 export * from './data.log.schema';
-export * from './user.schema';
 export * from './events.schema';
+export * from './user.schema';
 
 const SCHEMA_LIST = [
   { name: User.name, schema: UserSchema, dbPrefix: 'APP' },
@@ -42,14 +40,4 @@ export const SchemaProviders = SCHEMA_LIST.map((model) => ({
   inject: [`${model.name}Connection`],
 }));
 
-@Module({
-  imports: [
-    ConfigModule.forRoot({
-      isGlobal: true,
-      load: [configs],
-    }),
-  ],
-  providers: [...ConnectionProviders, ...SchemaProviders],
-  exports: [...ConnectionProviders, ...SchemaProviders, ConfigModule],
-} as any)
-export class DBModule {}
+export { DBModule } from './db.module';
